# Try to isolate debug mode perf regression