--- a/examples/debug_perf.rs
+++ b/examples/debug_perf.rs
@@ -1,17 +1,20 @@
 //! example for manually testing the perf of color-eyre in debug vs release
 
 use color_eyre::{Help, Report};
-use eyre::WrapErr;
-use tracing::{info, instrument};
+use eyre::{eyre, WrapErr};
+use tracing::instrument;
 
 #[instrument]
 fn main() -> Result<(), Report> {
     #[cfg(feature = "capture-spantrace")]
     install_tracing();
 
-    let report = read_config().unwrap_err();
-
     let start = std::time::Instant::now();
+    let report = Err::<(), Report>(eyre!("fake error"))
+        .wrap_err("wrapped error")
+        .suggestion("try using a file that exists next time")
+        .unwrap_err();
+
     let _ = format!("Error: {:?}", report);
     drop(report);
     let end = std::time::Instant::now();
@@ -37,17 +40,4 @@ fn install_tracing() {
         .with(fmt_layer)
         .with(ErrorLayer::default())
         .init();
-}
-
-#[instrument]
-fn read_file(path: &str) -> Result<(), Report> {
-    info!("Reading file");
-    Ok(std::fs::read_to_string(path).map(drop)?)
-}
-
-#[instrument]
-fn read_config() -> Result<(), Report> {
-    read_file("fake_file")
-        .wrap_err("Unable to read config")
-        .suggestion("try using a file that exists next time")
 }