fixes #19 

initial tests aren't reproducing the issue yet, we're seeing an 8x decrease in perf between debug and release but not 1000x and definitely not 3-5 second delays.

```
color-eyre/examples on  master [$?] via 🦀 v1.43.1 took 15s
❯ RUST_BACKTRACE=full cargo run --example debug_perf
   Compiling color-eyre v0.3.2 (/home/jlusby/git/rust/color-eyre)
    Finished dev [unoptimized + debuginfo] target(s) in 1.94s
     Running `/home/jlusby/git/rust/color-eyre/target/debug/examples/debug_perf`
May 25 10:46:20.784  INFO read_config:read_file{path="fake_file"}: Reading file
[examples/debug_perf.rs:19] end - start = 716.4µs

color-eyre/examples on  master [$?] via 🦀 v1.43.1 took 2s
❯ RUST_BACKTRACE=1 cargo run --example debug_perf
    Finished dev [unoptimized + debuginfo] target(s) in 0.03s
     Running `/home/jlusby/git/rust/color-eyre/target/debug/examples/debug_perf`
May 25 10:46:25.543  INFO read_config:read_file{path="fake_file"}: Reading file
[examples/debug_perf.rs:19] end - start = 552.331µs

color-eyre/examples on  master [$?] via 🦀 v1.43.1
❯ cargo run --example debug_perf
    Finished dev [unoptimized + debuginfo] target(s) in 0.04s
     Running `/home/jlusby/git/rust/color-eyre/target/debug/examples/debug_perf`
May 25 10:46:29.709  INFO read_config:read_file{path="fake_file"}: Reading file
[examples/debug_perf.rs:19] end - start = 154.044µs

color-eyre/examples on  master [$?] via 🦀 v1.43.1
❯ cargo run --example debug_perf --release
   Compiling color-eyre v0.3.2 (/home/jlusby/git/rust/color-eyre)
    Finished release [optimized] target(s) in 1.43s
     Running `/home/jlusby/git/rust/color-eyre/target/release/examples/debug_perf`
May 25 10:46:37.432  INFO read_config:read_file{path="fake_file"}: Reading file
[examples/debug_perf.rs:19] end - start = 24.35µs
```