@yaahc 2020-05-25
I updated the example to only create an error but to also include that construction time in its timing and I'm still not seeing the level of slowdown reported.

```
color-eyre/examples on  debug-perf [$!] via 🦀 v1.43.1 took 7s
❯ RUST_BACKTRACE=full cargo run --example debug_perf
   Compiling color-eyre v0.3.2 (/home/jlusby/git/rust/color-eyre)
    Finished dev [unoptimized + debuginfo] target(s) in 1.82s
     Running `/home/jlusby/git/rust/color-eyre/target/debug/examples/debug_perf`
[examples/debug_perf.rs:22] end - start = 329.155443ms

color-eyre/examples on  debug-perf [$!] via 🦀 v1.43.1
❯ RUST_BACKTRACE=full cargo run --example debug_perf --release
   Compiling color-eyre v0.3.2 (/home/jlusby/git/rust/color-eyre)
    Finished release [optimized] target(s) in 1.39s
     Running `/home/jlusby/git/rust/color-eyre/target/release/examples/debug_perf`
[examples/debug_perf.rs:22] end - start = 6.760893ms

color-eyre/examples on  debug-perf [$!] via 🦀 v1.43.1
❯ RUST_BACKTRACE=full cargo run --example debug_perf
    Finished dev [unoptimized + debuginfo] target(s) in 0.03s
     Running `/home/jlusby/git/rust/color-eyre/target/debug/examples/debug_perf`
[examples/debug_perf.rs:22] end - start = 328.418983ms

color-eyre/examples on  debug-perf [$!] via 🦀 v1.43.1
❯ RUST_BACKTRACE=1 cargo run --example debug_perf --release
    Finished release [optimized] target(s) in 0.03s
     Running `/home/jlusby/git/rust/color-eyre/target/release/examples/debug_perf`
[examples/debug_perf.rs:22] end - start = 7.007158ms

color-eyre/examples on  debug-perf [$!] via 🦀 v1.43.1
❯ RUST_BACKTRACE=1 cargo run --example debug_perf
    Finished dev [unoptimized + debuginfo] target(s) in 0.03s
     Running `/home/jlusby/git/rust/color-eyre/target/debug/examples/debug_perf`
[examples/debug_perf.rs:22] end - start = 328.957609ms

color-eyre/examples on  debug-perf [$!] via 🦀 v1.43.1
❯ cargo run --example debug_perf --release
    Finished release [optimized] target(s) in 0.03s
     Running `/home/jlusby/git/rust/color-eyre/target/release/examples/debug_perf`
[examples/debug_perf.rs:22] end - start = 13.412µs

color-eyre/examples on  debug-perf [$!] via 🦀 v1.43.1
❯ cargo run --example debug_perf
    Finished dev [unoptimized + debuginfo] target(s) in 0.03s
     Running `/home/jlusby/git/rust/color-eyre/target/debug/examples/debug_perf`
[examples/debug_perf.rs:22] end - start = 100.443µs
```