@yaahc 2020-05-25
This does show up to .3 seconds to create a report when backtraces are being captured in debug mode, which is slower than I ever realized it is but also this is why we emulated `std::backtrace::Backtrace` and made it not capture backtraces by default.