@yaahc 2020-05-25
Here's an extra  one without capturing a backtrace that prints the error to the screen instead of to a String. This one takes a bit longer but it also narrows the gap between debug and release.

```
color-eyre/examples on  debug-perf [$!] via 🦀 v1.43.1
❯ cargo run --example debug_perf
   Compiling color-eyre v0.3.2 (/home/jlusby/git/rust/color-eyre)
    Finished dev [unoptimized + debuginfo] target(s) in 1.86s
     Running `/home/jlusby/git/rust/color-eyre/target/debug/examples/debug_perf`
Error:
   0: wrapped error
   1: fake error

Suggestion: try using a file that exists next time
[examples/debug_perf.rs:22] end - start = 111.612µs

color-eyre/examples on  debug-perf [$!] via 🦀 v1.43.1
❯ cargo run --example debug_perf --release
   Compiling color-eyre v0.3.2 (/home/jlusby/git/rust/color-eyre)
    Finished release [optimized] target(s) in 1.38s
     Running `/home/jlusby/git/rust/color-eyre/target/release/examples/debug_perf`
Error:
   0: wrapped error
   1: fake error

Suggestion: try using a file that exists next time
[examples/debug_perf.rs:22] end - start = 28.32µs
```