@yaahc 2020-05-25
Adding some proper tracing instrumentation slows it down a little further, but again this is the same slowdown of 30us on both release and debug.

```
color-eyre/examples on  debug-perf [$!] via 🦀 v1.43.1 took 34s
❯ cargo run --example debug_perf --release
   Compiling color-eyre v0.3.2 (/home/jlusby/git/rust/color-eyre)
    Finished release [optimized] target(s) in 1.37s
     Running `/home/jlusby/git/rust/color-eyre/target/release/examples/debug_perf`
Error:
   0: wrapped error
   1: fake error

  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ SPANTRACE ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

   0: debug_perf::time_report
      at examples/debug_perf.rs:16

Suggestion: try using a file that exists next time
[examples/debug_perf.rs:28] end - start = 43.514µs

color-eyre/examples on  debug-perf [$!] via 🦀 v1.43.1
❯ cargo run --example debug_perf
   Compiling color-eyre v0.3.2 (/home/jlusby/git/rust/color-eyre)
    Finished dev [unoptimized + debuginfo] target(s) in 1.88s
     Running `/home/jlusby/git/rust/color-eyre/target/debug/examples/debug_perf`
Error:
   0: wrapped error
   1: fake error

  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ SPANTRACE ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

   0: debug_perf::time_report
      at examples/debug_perf.rs:16

Suggestion: try using a file that exists next time
[examples/debug_perf.rs:28] end - start = 150.204µs
```