@yaahc 2020-05-25
Adding extra tracing spans seems to increase the cost for debug builds but not for release builds, so that might be a lead as to what is going on.

```
color-eyre/examples on  debug-perf [$!] via 🦀 v1.43.1 took 24s
❯ cargo run --example debug_perf
   Compiling color-eyre v0.3.2 (/home/jlusby/git/rust/color-eyre)
    Finished dev [unoptimized + debuginfo] target(s) in 1.90s
     Running `/home/jlusby/git/rust/color-eyre/target/debug/examples/debug_perf`
Error:
   0: wrapped error
   1: fake error

  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ SPANTRACE ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

   0: debug_perf::time_report_inner
      at examples/debug_perf.rs:21
   1: debug_perf::time_report
      at examples/debug_perf.rs:16

Suggestion: try using a file that exists next time
[examples/debug_perf.rs:33] end - start = 173.273µs

color-eyre/examples on  debug-perf [$!] via 🦀 v1.43.1 took 2s
❯ cargo run --example debug_perf --release
   Compiling color-eyre v0.3.2 (/home/jlusby/git/rust/color-eyre)
    Finished release [optimized] target(s) in 1.38s
     Running `/home/jlusby/git/rust/color-eyre/target/release/examples/debug_perf`
Error:
   0: wrapped error
   1: fake error

  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ SPANTRACE ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

   0: debug_perf::time_report_inner
      at examples/debug_perf.rs:21
   1: debug_perf::time_report
      at examples/debug_perf.rs:16

Suggestion: try using a file that exists next time
[examples/debug_perf.rs:33] end - start = 43.583µs
```