@yaahc 2020-05-25
Further testing makes me believe that the likely culprit is that `backtrace::Backtrace` doesn't get optimizations in debug mode where as `std::backtrace::Backtrace` does thanks to it being precompiled as part of std. For now I've updated the documentation to indicate this issue and how to resolve it, I'm going to stop investigating on the assumption that this is the problem unless I get a response saying that the issue still exists.